feat: add projections refactoring
