+++ src/main/java/it/pagopa/transactions/projections/TransactionsProjection.java
@@ -1,0 +1,15 @@
+package it.pagopa.transactions.projections;
+
+import it.pagopa.transactions.model.RptId;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TransactionsProjection<T> {
+    
+    private RptId rptId;
+    private T data;
+}
