+++ src/main/java/it/pagopa/transactions/projections/handlers/ProjectionHandler.java
@@ -1,0 +1,6 @@
+package it.pagopa.transactions.projections.handlers;
+
+public interface ProjectionHandler<T, S> {
+
+  S handle(T data);
+}
